requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
jinja2==3.1.4
python-multipart==0.0.12
pyyaml==6.0.2
